refactor cpp compiler type inference
